Problem: need basic server design
Solution: faster than the eye can see, Hintjens rips off the server core
from ZCCP and thunks that into the Malamute Protocol. The audience doesn't
quite follow, which is the secret to every magic trick. In about an hour
and a quarter, we get from protocol wireframe to a working stream server
that can do transient pubsub using regular expression matching.

If you want to look at how it's done, the key is mlm_server.xml, which is
the state machine model for the server. This defines the logic for each
message, according to the current client connection state.

If this is easy, it's because we're standing on the rather bruised and yet
surprisingly solid shoulders of zproto and its antecedants.