Problem: mshell does not reliably send its message
This is due to using dual pipes; the message is still being sent
on the data pipe when the program shuts down with synchronization
on the command pipe. The proper solution is to implement trackers
so mshell can wait for a delivery confirmation before exiting.

Solution: pause for 200msec to give delivery time to complete.

Fixes #35