Problem: mlm_msg reference counting is not thread safe
Solution: use lockfree reference counting. To make this work in a
portable fashion, I extended the libzmq API, so Malamute needs the
libzmq master version of 2014/12/23 or later. This is marked as
version 4.2.0.

The code is now theoretically threadsafe in that a stream engine
can send the same message to many other threads, which can then
drop their references as they like.