Problem: we lose mailbox messages when reconnecting with same address
The problem hits after N second, depending on the timeout in malamute.cfg.
What happens is that an existing old client thread wakes up, decides that
it's expired, and kills the client mailbox. The current active client
loses any messages waiting in the mailbox.

Solution: when registering a new client, force an expiration of any
existing client using the same address. This is the only way to safely
reconnect after an unexpected disconnect. If we reject new clients with
the same / existing address, we cannot reconnect until the timeout has
expired.

Fixes #53