Problem: Can't run make check
Running make check on the latest master gets stuck. Culprit from the
bisect point of view seems to be commit c139362. In the fact, in the
light of this commit test seems to be4 stuck kinda correctly as two
clients shouldn't have the same name.

Solution: Use two different client names in the test