Problem: malamute won't start without a config file
This is annoying when trying to run from scratch.

Solution: the malamute.c main program handles this situation by
creating a new config file with default values, and reports this
to the user.

It saves this file to disk so the second time it runs, it'll use
it. The user can then edit the file. This is simpler than copying
a file from elsewhere.