Problem: mlm_client cannot recover from disconnected state. Solution: Instead of ignoring any heartbeats turn off the timer.
When the client entered disconnected stated the heartbeat event is
overridden to prevent unnecessary PING messages to be send to a dead
server. This also resulted in ignoring the heartbeat events from a
revived server which resulted in a client stuck in disconnected state.
Instead of overriding the heartbeat event we stop the heartbeat timer
now. This prevents sending any unnecessary PING messages to a dead
server but one the same time retains the ability to listen to heartbeat
messages from a revived client.

I also modified the tests to cover this use case.