Problem: server can segfault due to incomplete connections
(trying again -- the tests pass for me.  I'm not sure why they failed
before -- perhaps they have spurious failures?).

If the server is restarted and a client is still connected, it may
have a few CONNECTION_PING messages queued up on the socket which the
server will receive when it is restarted.  This is fine at the moment,
but in the case where the client does not then send CONNECTION_OPEN,
the server will try to deregister the client, and since the address
was never set, it segfaults.

Maybe this is only something I see because I'm working on a python
client and it's not well-behaved or something, but the server
shouldn't crash regardless, because even with well-behaved clients
this could happen if the client shuts down just after the server is
restarted.

Solution: cleanly handle the case of the client address being null.
Add a regression test.