Problem: want to play with meson build system
Solution: write a basic meson.build file for malamute by hand, note
number of issues and push it upstream.

meson build is a generator (like CMake or zproject) of build recipes. It
uses ninja backend by default, but have an ability to generate builds
for MSVC and XCode. Like zproject it's quite opiniated and declarative
(unline WAF) and easy to write (unline CMake), so can be a great
match.

How to use:
    mkdir build
    meson . build/
    ninja -v -C build
    ninja -v -C build test

The final goal is to get zproject_meson.gsl, so this will give people
option to play with it.