Properly handle pending stream traffic after client disconnect
The current code inserts a one second delay in the hope that any pending
stream traffic will be flushed before the client context is destroyed.
Under heavy load, this may not work and we end up with a use-after-free
in s_forward_stream_traffic(). Instead of relying on luck, do proper
reference counting on the client contexts.